OperationsInterface created, Withdraw operation implmeneted and balance type changed to double

+++ BankApp-JavaCore/src/main/java/Project/DB/DAO.java
@@ -11,6 +11,6 @@ public interface DAO {
 
     public UserEntity getUserById(String id);
 
-    public void updateBalance(String id, int newBalance);
+    public void updateBalance(String id, double newBalance);
 
 }
