+++ BankApp-JavaCore/src/main/java/Project/DB/DaoImpl.java
@@ -13,20 +13,20 @@ public class DaoImpl implements DAO {
 
         String sql = "INSERT INTO userdata (id, username,password,name,surname,age,address,balance) VALUES (?,?,?,?,?,?,?,?)";
 
-        try(Connection conn = DbConnection.getConnection();
-            PreparedStatement stmt = conn.prepareStatement(sql)){
-            stmt.setString(1,idGenerator.generateId());
-            stmt.setString(2,user.getUsername());
-            stmt.setString(3,user.getPassword());
-            stmt.setString(4,user.getName());
-            stmt.setString(5,user.getSurname());
-            stmt.setInt(6,user.getAge());
-            stmt.setString(7,user.getAddress());
-            stmt.setInt(8,user.getBalance());
+        try (Connection conn = DbConnection.getConnection();
+             PreparedStatement stmt = conn.prepareStatement(sql)) {
+            stmt.setString(1, idGenerator.generateId());
+            stmt.setString(2, user.getUsername());
+            stmt.setString(3, user.getPassword());
+            stmt.setString(4, user.getName());
+            stmt.setString(5, user.getSurname());
+            stmt.setInt(6, user.getAge());
+            stmt.setString(7, user.getAddress());
+            stmt.setDouble(8, user.getBalance());
 
             stmt.executeUpdate();
 
-        }catch (SQLException e) {
+        } catch (SQLException e) {
             e.printStackTrace();
         }
         System.out.println("Successfully Registered");
@@ -37,11 +37,11 @@ public class DaoImpl implements DAO {
     public UserEntity getUser(String userName, String password) {
         String sql = "SELECT * FROM userdata WHERE username=? AND password=?";
 
-        try(Connection conn = DbConnection.getConnection();
-            PreparedStatement stmt = conn.prepareStatement(sql)){
+        try (Connection conn = DbConnection.getConnection();
+             PreparedStatement stmt = conn.prepareStatement(sql)) {
 
-            stmt.setString(1,userName);
-            stmt.setString(2,password);
+            stmt.setString(1, userName);
+            stmt.setString(2, password);
             ResultSet rs = stmt.executeQuery();
 
             if (rs.next()) {
@@ -56,7 +56,7 @@ public class DaoImpl implements DAO {
                         rs.getInt("balance")
                 );
             }
-        }catch (SQLException e){
+        } catch (SQLException e) {
             e.printStackTrace();
         }
         return null;
@@ -68,10 +68,10 @@ public class DaoImpl implements DAO {
     public UserEntity getUserById(String id) {
         String sql = "SELECT * FROM userdata WHERE id=?";
 
-        try(Connection conn = DbConnection.getConnection();
-            PreparedStatement stmt = conn.prepareStatement(sql)){
+        try (Connection conn = DbConnection.getConnection();
+             PreparedStatement stmt = conn.prepareStatement(sql)) {
 
-            stmt.setString(1,id);
+            stmt.setString(1, id);
 
             ResultSet rs = stmt.executeQuery();
 
@@ -84,22 +84,22 @@ public class DaoImpl implements DAO {
                         rs.getString("password"),
                         rs.getInt("age"),
                         rs.getString("address"),
-                        rs.getInt("balance")
+                        rs.getDouble("balance")
                 );
             }
-        }catch (SQLException e){
+        } catch (SQLException e) {
             e.printStackTrace();
         }
         return null;
     }
 
     @Override
-    public void updateBalance(String id, int newBalance) {
+    public void updateBalance(String id, double newBalance) {
         String sql = "UPDATE userdata SET balance = ? WHERE id = ?";
         try (Connection conn = DbConnection.getConnection();
              PreparedStatement stmt = conn.prepareStatement(sql)) {
 
-            stmt.setInt(1, newBalance);
+            stmt.setDouble(1, newBalance);
             stmt.setString(2, id);
 
             stmt.executeUpdate();
